+++ apps/desktop/lib/services/noop_updater.dart
@@ -2,7 +2,7 @@ import 'package:remote_rift_updater/remote_rift_updater.dart';
 
 class NoopUpdater implements ApplicationUpdater {
   @override
-  Future<void> acknowledgeHealthyStart() async {}
+  Future<UpdateStartupResult> acknowledgeHealthyStart() async => NoUpdate();
 
   @override
   Future<UpdateRelease?> checkUpdateAvailable() async => null;