+++ packages/ui/lib/remote_rift_ui.dart
@@ -7,4 +7,5 @@ export 'src/theme/theme.dart';
 export 'src/theme/theme_extension.dart';
 export 'src/theme/theme_types.dart';
 export 'src/widgets/lifecycle.dart';
+export 'src/widgets/events_listener.dart';
 export 'src/widgets/app_version.dart';